[Generator] Refactor NamespaceManager to its own location.
Move the class to its own file to simplify changes and code review.
Enable nullability in the class.
